pyproject.toml
[project]
name = "test-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "bs4>=0.0.2",
    "discord-py[voice]>=2.7.0",
    "fastapi>=0.135.2",
    "flask>=3.1.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.42.0",
    "xmltodict>=1.0.4",
]
